+++ app/assets/javascripts/application.js
@@ -13,11 +13,9 @@
 
 //= require application/lib/jquery-1.12.0.min
 //= require application/lib/underscore-min
-//= require application/lib/bootstrap
 
 //= require turbolinks
 
 //= require application/lib/app
-//= require application/lib/backbone
 
 //= require bootstrap
